rename: 사용자 권한 enum 주소 변경
